corrected some errors in wc

+++ Kernel/include/pipeManager.h
@@ -15,8 +15,8 @@ PipeManagerADT create_pipe_manager();
 int16_t get_pipe_fd();
 int16_t open_pipe(int id, char mode);
 int16_t close_pipe(uint16_t fd);
-int16_t write_pipe(uint16_t fd, char *buffer, uint16_t *count);
-int16_t read_pipe(uint16_t fd, char *buffer, uint16_t *count);
+int32_t write_pipe(uint16_t fd, char *buffer, uint32_t *count);
+int32_t read_pipe(uint16_t fd, char *buffer, uint32_t *count);
  
 
 #endif
